host-retriever/requirements.txt
# DSA Knowledge Base (from backend)
qdrant-client>=1.7.0
openai>=1.0.0
httpx>=0.25.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# API
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

